ku1255_sim: Fix key matrix impedance computation.
Implement N-sources voltage divider.
Use the correct equivalent parallel resistance equation and not the
special-case for when all resistors are equal.
Include column pull-ups in the computation (should not change the result,
but they are there in reality, so why not).
Also, fix a brown-paper-bag mistake: raising would have worked better with
an actual "raise" statement...