simsn8: Fix watchdog periodicity.
Watchdog triggers when 12th bit would be set, making it a 11 bits counter.
Documentation mentions 8192, which would be a 13-bits counter, but that
does not match with documented overflow periods. Assume documented periods
are correct. Also, the formula given in the spec is incorrect: dividing
8k by slow clock period in seconds gives very large values (24 and 49
million seconds). Either it is multiplying which makes sense, or the
right-hand-side unit should be the frequency and not the period.
Also, subtract slow_clock_threshold from current value instead of setting
it to zero, to preserve sub-slow-clock-cycle accuracy: otherwise the error
would be multiplied by 2**11, which is enough to deviate from spec by a
few milliseconds.