asmlib/usb.asm: Fix SETUP-to-DATA transition.
Deferred stall seems meaningless on this µC: the SETUP and 8 following DATA
OUT bytes containing the setup fields are already ACK'ed  (under control
of FEP0SETUP), so by the time any function runs we are already in the DATA
stage of the transaction.
So:
- for control write OUT transfers (wLength * OUT), we can either ACK to let
  the µC know we are done processing the SETUP fields and it can overwrite
  the content of the EP0 buffer with whatever the host is sending in the
  upcoming DATA OUT packets, or we can STALL so these DATA OUT get
  rejected. Such transfer may be the IN handshake in case of wLength == 0.
- for control read IN transfers, we can either put data in the EP0 buffer
  and ACK to let the µC send this chunk of data to the host, or STALL so
  these DATA IN get rejected.
As a consequence:
- call usb_stall_ep0 instead of usb_deferred_stall_ep0
- drop _ep0_stall_next_stage
- make usb_ack_ep0 take the number of bytes written to buffer from register
  A, and update callers to set it
- jump to _handle_ep0_in at the end of GET_DESCRIPTOR handler, to write the
  first min(wLength, 8) bytes to EP0 buffer.
Also, clear FEP0{SETUP,OUT,IN} in usb_stall_ep0 and remove the last few
scattered instructions clearing these bits.