asmlib/usb.asm: Test wLengthH equality before inferiority.
As a tiny optimisation, as most descriptors are less than 256 bytes long
(including sequences of descriptors, as in the case of configuration
descriptors), so both wLengthH and _usb_setup_data_len_m are likely to be
equal (by both being zero). Moreover, it is likely that the host knows the
descriptor length as well, in which case this optimisation also applies.
Also, add a few comments.