sn8.simsn8.SN8: Optimise internal OSCM and FGIE accesses.
A slight downside is that these accesses do not trigger read watchers
anymore, but triggering watchers for such internal accesses was a bit of a
gray zone: these accesses are not really under firmware control, so
triggering read watchers here is unlikely to be helpful.
Also, fix a comment typo.